session handling to rpc.go
