+++ service/tbc/tbc.go
@@ -1230,41 +1230,6 @@ func (s *Server) AddressBalance(ctx context.Context, encodedAddress string) (uin
 	return balance, nil
 }
 
-func (s *Server) newSession(ws *tbcWs) (string, error) {
-	b := make([]byte, 16)
-
-	for {
-		// Create random hexadecimal string to use as an ID
-		_, err := rand.Read(b)
-		if err != nil {
-			return "", err
-		}
-		id := hex.EncodeToString(b)
-
-		// Ensure the key is not already in use, if it is then try again.
-		s.mtx.Lock()
-		if _, ok := s.sessions[id]; ok {
-			s.mtx.Unlock()
-			continue
-		}
-		s.sessions[id] = ws
-		s.mtx.Unlock()
-
-		return id, nil
-	}
-}
-
-func (s *Server) deleteSession(id string) {
-	s.mtx.Lock()
-	_, ok := s.sessions[id]
-	delete(s.sessions, id)
-	s.mtx.Unlock()
-
-	if !ok {
-		log.Errorf("id not found in sessions %s", id)
-	}
-}
-
 // DBOpen opens the undelying server database. It has been put in its own
 // function to make it available during tests and hemictl.
 func (s *Server) DBOpen(ctx context.Context) error {
